refactor(admin-console): extract AuditLogTable from AuditLog.tsx
AuditLog.tsx rendered the filter form and the results table in one 367-line
page, pulling 17 imports (over the high-coupling threshold of 16). The results
table owned the only Table / StatusIndicator / Icon / Box uses and the
formatRelativeTime helper (plus the outcomeIndicator badge logic). Extracting it
to AuditLogTable.tsx drops AuditLog.tsx to 13 imports (off the high-coupling
list) and gives the table its own single-responsibility module.

Relates #1418.

## Regression analysis
- Pure structural move: the table JSX + outcomeIndicator are byte-identical,
  relocated and rendered via <AuditLogTable items={items} loading={loading} />.
  admin-console typecheck passes; all 143 tests pass unchanged.

## Physical impact
- NO-OP on CFn / deployed artifacts. Frontend source reorganization only.
